refactor: split policy canonical/numeric files; count SSE drops in Send
