+++ internal/api/stream.go
@@ -81,7 +81,7 @@ func (h *StreamHandler) Handle(w http.ResponseWriter, r *http.Request) {
 	// Register for live events before gap-fill so events arriving during replay
 	// buffer in the subscriber queue instead of being missed (an overlap with
 	// replay yields duplicates, deduped client-side by id: — at-least-once).
-	sub := stream.NewSubscriber(claims)
+	sub := stream.NewSubscriber(claims, h.Metrics)
 	h.Hub.Add(topic, role, sub)
 	defer h.Hub.Remove(topic, role, sub)
