+++ internal/policy/policy.go
@@ -1,12 +1,8 @@
 package policy
 
 import (
-	"encoding/json"
 	"fmt"
-	"math"
-	"math/big"
 	"regexp"
-	"strconv"
 	"strings"
 
 	"github.com/Wave-RF/WaveHouse/internal/chsql"
@@ -419,214 +415,6 @@ func resolveInValues(tmpl string, claims map[string]any) []any {
 	// nil (not a one-element set containing the partial literal), so it fails
 	// closed like the bare-claim case above (#385).
 	return nil
-}
-
-// maxCanonicalDigits bounds both a numeric literal's digit count and its
-// exact decimal expansion. Big-integer parsing is superlinear in digit count
-// and the ingest check path hands CanonicalScalar client-controlled literals
-// (CWE-400), and a short exponent literal can hide a wide expansion ("1e-150"
-// is six characters with a 152-character exact form). 100 digits is far past
-// any real id — uint256 is 78.
-const maxCanonicalDigits = 100
-
-// CanonicalScalar renders a decoded JSON value as the canonical string the
-// policy layer binds and compares, reporting ok=false for values with no such
-// form: null, objects, and arrays. A structured value is never a sensible
-// scalar comparison value — it usually means a dropped path segment
-// ({{ jwt.app_metadata }} for {{ jwt.app_metadata.tenant_id }}), and binding
-// fmt.Sprint's "map[…]"/"[…]" rendering would let _neq/_lt match essentially
-// every row; the one legitimate structured shape, a bare-claim _in array, is
-// unpacked by resolveInValues before its elements reach here. A json.Number
-// (jwt.WithJSONNumber on claims, UseNumber on ingest payloads) binds in
-// canonical decimal form, not the token's spelling: "1", "1.0", and "1e3" are
-// one JSON value, and a numeric ClickHouse column rejects '1.0'/'1e3' as a
-// per-query TYPE_MISMATCH error. The canonical form is exact at every width
-// and precision — integer literals via big.Int, fractions and exponents via
-// canonicalDecimal, never a float64 round-trip that could bind a value the
-// token doesn't carry ("1e-400" fails closed rather than collapsing to "0").
-// A literal, or an exact form, past maxCanonicalDigits likewise has no
-// canonical form and fails closed (1e400, 1e-400). Claim resolution and the
-// insert-check comparison's two sides (internal/api) all route through this
-// one function, so what a read filter binds and what a write check accepts
-// can't drift.
-func CanonicalScalar(v any) (string, bool) {
-	switch val := v.(type) {
-	case nil, map[string]any, []any:
-		return "", false
-	case json.Number:
-		// The digit bound guards the superlinear big.Int parse on the
-		// client-controlled ingest path. It counts digits, not bytes — a sign,
-		// point, or exponent marker doesn't feed the big parse — so "-1e99"
-		// binds exactly like its written-out form, matching the "one JSON
-		// value" contract below (only at the bound's very edge can the exact-form
-		// gate's slack separate two spellings). canonicalDecimal re-checks its
-		// own exact-form length, where a short literal can hide a wide expansion.
-		digits := 0
-		for i := 0; i < len(val); i++ {
-			if val[i] >= '0' && val[i] <= '9' {
-				digits++
-			}
-		}
-		if digits > maxCanonicalDigits {
-			return "", false
-		}
-		if i, ok := new(big.Int).SetString(val.String(), 10); ok {
-			return i.String(), true
-		}
-		return canonicalDecimal(val.String())
-	case float64:
-		// Production claims arrive as json.Number (jwt.WithJSONNumber), but
-		// Evaluate is an exported API and float64 is what a plain json.Unmarshal
-		// hands any other caller. At or past 2^53 the decode has already
-		// collapsed neighboring JSON integers onto one float — ANY digits
-		// rendered for it could be another principal's ID — so refuse, in depth
-		// (#381 review). Below that, render positionally: fmt.Sprint's exponent
-		// form ("1e+06") is a spelling numeric ClickHouse columns reject.
-		if math.Abs(val) >= 1<<53 {
-			return "", false
-		}
-		return strconv.FormatFloat(val, 'f', -1, 64), true
-	default:
-		return fmt.Sprint(val), true
-	}
-}
-
-// LiteralValue marks an insert-check required value the policy author wrote
-// as a placeholder-free literal (Evaluate). A literal carries no JSON type —
-// "1.0" means the number 1 to a numeric column and the three-character text
-// to a String column — so the check comparison (internal/api) accepts its
-// numeric reading as well as its spelling. The type is the gate: a
-// claim-derived value is never wrapped, so a string-typed claim keeps strict
-// canonical equality and can't gain a numeric reading it didn't have. Only
-// CheckClauses carries this type; read filters bind plain strings.
-type LiteralValue string
-
-// CanonicalNumericLiteral renders a policy-authored literal that spells a
-// JSON number in canonical decimal form ("1.0" → "1"), reporting ok=false for
-// everything else. The json.Valid gate keeps this to spellings JSON itself
-// can produce: big.Int would also take "+5" or "007", readings no decoded
-// claim or payload value ever has. It canonicalizes nothing at resolve time —
-// the literal still binds and auto-injects exactly as written; only the check
-// comparison consults this second reading.
-func CanonicalNumericLiteral(s string) (string, bool) {
-	if !json.Valid([]byte(s)) {
-		return "", false
-	}
-	return CanonicalScalar(json.Number(s))
-}
-
-// canonicalDecimal renders a non-integer JSON number literal (one carrying a
-// fraction or exponent) as its exact canonical decimal string: "1.0" → "1",
-// "2.50" → "2.5", "1e3" → "1000", "25e-4" → "0.0025", every digit preserved
-// at any precision. ok=false for anything that is not a JSON number and for
-// any literal whose exact decimal form would exceed maxCanonicalDigits.
-// Exactness is the point: rounding through float64 would collapse "1e-400"
-// to "0" and land wide decimals on their neighbors, either way binding a
-// value the token doesn't carry.
-func canonicalDecimal(lit string) (string, bool) {
-	mant, expStr, hasExp := strings.Cut(lit, "e")
-	if !hasExp {
-		mant, expStr, hasExp = strings.Cut(lit, "E")
-	}
-	exp := 0
-	if hasExp {
-		var err error
-		// Atoi rejects an empty or non-numeric exponent. The magnitude bound is
-		// the allocation guard, not a redundancy: the exact-form length check at
-		// the bottom runs only after strings.Repeat has already built the
-		// string, so without this bound a 12-byte "1e1000000000" would allocate
-		// a ~1 GiB expansion before being rejected.
-		if exp, err = strconv.Atoi(expStr); err != nil || exp > 2*maxCanonicalDigits || exp < -2*maxCanonicalDigits {
-			return "", false
-		}
-	}
-	sign := ""
-	if strings.HasPrefix(mant, "-") {
-		sign, mant = "-", mant[1:]
-	}
-	intPart, fracPart, _ := strings.Cut(mant, ".")
-	digits := intPart + fracPart
-	if digits == "" {
-		return "", false
-	}
-	for i := 0; i < len(digits); i++ {
-		if digits[i] < '0' || digits[i] > '9' {
-			return "", false
-		}
-	}
-	// point is where the decimal point sits in digits once the exponent is
-	// applied: digits[:point] to its left, digits[point:] to its right.
-	point := len(intPart) + exp
-	for len(digits) > 0 && digits[0] == '0' {
-		digits = digits[1:]
-		point--
-	}
-	for len(digits) > 0 && len(digits) > point && digits[len(digits)-1] == '0' {
-		digits = digits[:len(digits)-1]
-	}
-	if digits == "" {
-		// Every spelling of zero ("0.0", "-0e9") is the one value 0 —
-		// signless, since a numeric column can't parse "-0".
-		return "0", true
-	}
-	var out string
-	switch {
-	case point <= 0:
-		out = "0." + strings.Repeat("0", -point) + digits
-	case point >= len(digits):
-		out = digits + strings.Repeat("0", point-len(digits))
-	default:
-		out = digits[:point] + "." + digits[point:]
-	}
-	// The exact form is what must stay small: "1e-150" is a six-character
-	// literal with a 152-character expansion. The +2 slack exists for a "0."
-	// prefix, though any form may spend it (a 102-digit integer expansion
-	// passes). This runs after the Repeat above, so it bounds what callers
-	// see — the exponent bound is what keeps the allocation itself small.
-	if len(out) > maxCanonicalDigits+2 {
-		return "", false
-	}
-	return sign + out, true
-}
-
-// compareCanonicalDecimals orders two canonical decimal forms (CanonicalScalar
-// output) as numbers, by digit-string arithmetic alone — the comparison twin of
-// canonicalDecimal, sharing its invariants: an optional leading '-' (never on
-// zero), no leading integer zeros except a lone "0", no trailing fraction
-// zeros, no exponent. Those invariants are what make the string operations
-// sound: with no leading zeros a longer integer part IS the larger magnitude,
-// and with no trailing zeros a fraction that is a proper prefix of another IS
-// the smaller. Never a float round-trip, so 64-bit-plus IDs order exactly.
-func compareCanonicalDecimals(a, b string) int {
-	if a == b {
-		return 0
-	}
-	na, nb := strings.HasPrefix(a, "-"), strings.HasPrefix(b, "-")
-	switch {
-	case na && !nb:
-		return -1
-	case !na && nb:
-		return 1
-	case na && nb:
-		return -compareCanonicalMagnitudes(a[1:], b[1:])
-	}
-	return compareCanonicalMagnitudes(a, b)
-}
-
-// compareCanonicalMagnitudes orders two unsigned canonical forms.
-func compareCanonicalMagnitudes(a, b string) int {
-	ai, af, _ := strings.Cut(a, ".")
-	bi, bf, _ := strings.Cut(b, ".")
-	if len(ai) != len(bi) {
-		if len(ai) < len(bi) {
-			return -1
-		}
-		return 1
-	}
-	if c := strings.Compare(ai, bi); c != 0 {
-		return c
-	}
-	return strings.Compare(af, bf)
 }
 
 // navigateClaims traverses nested claim maps using dot-separated path parts.
