+++ internal/policy/rowfilter.go
@@ -2,8 +2,6 @@ package policy
 
 import (
 	"encoding/json"
-	"math/big"
-	"strconv"
 	"strings"
 	"time"
 )
@@ -17,221 +15,12 @@ func (p *ResolvedPermissions) HasRowFilter() bool {
 	return p != nil && len(p.rowFilter) > 0
 }
 
-// maxNumericOperandChars is the O(1) length pre-gate on numeric comparison
-// operands, checked before ANY scan of the value. It is verdict-preserving: a
-// JSON number literal carries at most four non-digit bytes (a sign, a decimal
-// point, an exponent marker and its sign), so anything longer already fails
-// the canonical digit bound (maxCanonicalDigits) — but proving that inside the
-// canonical gate costs a full json.Valid pass plus a digit count over a
-// client-controlled value, per subscriber per event on the fan-out goroutine.
-// The gate refuses first, without reading the bytes.
-const maxNumericOperandChars = maxCanonicalDigits + 4
-
 // maxTimeOperandChars is the same O(1) pre-gate for timestamp operands: the
 // ingest grammar's longest accepted spelling (RFC 3339 with nanoseconds and a
 // numeric offset) is 35 bytes, so 64 is generous slack — and the parser scans
 // its input, which without the gate a megabyte "timestamp" would make a
 // per-subscriber-per-event cost.
 const maxTimeOperandChars = 64
-
-// NumericFamily classifies how a ClickHouse numeric column stores a value —
-// the narrowing the row-filter comparison must apply to BOTH operands so its
-// verdict matches the query path, where ClickHouse narrows the stored value at
-// insert AND the filter constant at compare. The zero value is NumericNone: no
-// storage model, every comparison refused — the same fail-closed zero-value
-// contract as ColumnOpaque, so a future numeric type nobody classified can
-// never be compared under the wrong model.
-type NumericFamily uint8
-
-const (
-	NumericNone    NumericFamily = iota // unclassified: refuse, fail closed
-	NumericInteger                      // Int*/UInt*: exact at any width
-	NumericFloat                        // Float32/Float64: IEEE rounding at Bits
-	NumericDecimal                      // Decimal*: truncation at Scale
-)
-
-// NumericSpec is a numeric column's storage model. Bits is the bit width
-// (float width for NumericFloat, integer width for NumericInteger); Unsigned
-// marks UInt* (NumericInteger only); Precision and Scale are the stored total
-// and fractional digit counts (NumericDecimal only, 1 ≤ Precision ≤ 76).
-type NumericSpec struct {
-	Family    NumericFamily
-	Bits      int
-	Unsigned  bool
-	Precision int
-	Scale     int
-}
-
-// intBounds holds each ClickHouse integer width's inclusive decimal bounds as
-// canonical-form strings, computed once. A constant outside the width is not
-// reliably modelable — on one and the same release, ClickHouse was measured to
-// ERROR the comparison (a negative literal against an unsigned column: the
-// role reads no rows), to PROMOTE and compare mathematically ('256' against a
-// UInt8), and to WRAP at a width boundary ('9223372036854775808' against an
-// Int64 compares as −2^63, where exact-precision comparison would ADMIT the
-// −2^63 rows SQL hides under !=). Refusing out-of-range operands is the one
-// rule safe under all three behaviors; the cost is availability on bounds no
-// in-range data could ever satisfy differently.
-var intBounds = func() map[int]struct{ sMin, sMax, uMax string } {
-	m := make(map[int]struct{ sMin, sMax, uMax string }, 6)
-	one := big.NewInt(1)
-	for _, bits := range []int{8, 16, 32, 64, 128, 256} {
-		uMax := new(big.Int).Sub(new(big.Int).Lsh(one, uint(bits)), one)
-		sMax := new(big.Int).Sub(new(big.Int).Lsh(one, uint(bits-1)), one)
-		sMin := new(big.Int).Neg(new(big.Int).Lsh(one, uint(bits-1)))
-		m[bits] = struct{ sMin, sMax, uMax string }{sMin.String(), sMax.String(), uMax.String()}
-	}
-	return m
-}()
-
-// integerInRange reports whether a canonical integer form lies within the
-// column's width. An unknown width refuses — fail closed, never a guessed
-// range.
-func (n NumericSpec) integerInRange(c string) bool {
-	b, ok := intBounds[n.Bits]
-	if !ok {
-		return false
-	}
-	if n.Unsigned {
-		return compareCanonicalDecimals(c, "0") >= 0 && compareCanonicalDecimals(c, b.uMax) <= 0
-	}
-	return compareCanonicalDecimals(c, b.sMin) >= 0 && compareCanonicalDecimals(c, b.sMax) <= 0
-}
-
-// decimalInPrecision reports whether a canonical form's integer digits fit the
-// column's Precision−Scale budget. A payload past it is never storable
-// (DECIMAL_OVERFLOW rejects the insert); a constant past it was measured to
-// promote and compare mathematically on the query path — but the integer
-// widths' wrap behavior (intBounds) shows the same class is not reliably
-// modelable across pairs, so the refusal keeps one rule for every family at an
-// availability-only cost. A lone "0" integer part spends no digits
-// (Decimal(2,2) legally stores 0.99).
-func (n NumericSpec) decimalInPrecision(c string) bool {
-	if n.Precision < 1 || n.Scale < 0 || n.Scale > n.Precision {
-		// No coherent model: refuse, fail closed. The Scale bounds also protect
-		// truncateScale's slicing — a hand-built spec must degrade to refusal,
-		// never a panic on the fan-out goroutine.
-		return false
-	}
-	intPart, _, _ := strings.Cut(strings.TrimPrefix(c, "-"), ".")
-	digits := len(intPart)
-	if intPart == "0" {
-		digits = 0
-	}
-	return digits <= n.Precision-n.Scale
-}
-
-// compare orders two canonical decimal operands (numericCanonical /
-// CanonicalNumericLiteral output) in the column's storage domain, ok=false when
-// the model refuses the pair. Narrowing BOTH sides is what ClickHouse itself
-// does — it narrows the payload at insert and the bound constant at compare
-// (verified: Float32 stores 16777217 as 16777216 and `= '16777217'` still
-// matches; Decimal(10,2) stores 1.005 as 1.00 and `= '1.005'` still matches) —
-// so a threshold filter can no longer admit an event whose stored row lands on
-// the other side of the comparison (the ordering fail-open raised on #381).
-func (n NumericSpec) compare(a, b string) (int, bool) {
-	switch n.Family {
-	case NumericInteger:
-		// A fractional CONSTANT against an integer column is a per-query type
-		// error on the SQL path (the role reads no rows, loudly); a fractional
-		// PAYLOAD was never storable in the column. Refuse both — fail closed.
-		if strings.Contains(a, ".") || strings.Contains(b, ".") {
-			return 0, false
-		}
-		// Same rule for the column's range: ClickHouse's reading of an
-		// out-of-range constant varies by pair (error, mathematical promotion,
-		// or a width-boundary wrap that compares against a DIFFERENT value
-		// than written — see intBounds), and an out-of-range payload was never
-		// storable. Refuse both sides rather than model any one behavior.
-		if !n.integerInRange(a) || !n.integerInRange(b) {
-			return 0, false
-		}
-		return compareCanonicalDecimals(a, b), true
-	case NumericFloat:
-		fa, ok := narrowFloat(a, n.Bits)
-		if !ok {
-			return 0, false
-		}
-		fb, ok := narrowFloat(b, n.Bits)
-		if !ok {
-			return 0, false
-		}
-		// Both operands are now exact values of the column's float domain, so
-		// direct comparison IS the domain comparison — no ties left to break.
-		switch {
-		case fa < fb:
-			return -1, true
-		case fa > fb:
-			return 1, true
-		default:
-			return 0, true
-		}
-	case NumericDecimal:
-		// Precision is the range gate of the decimal family: a payload with
-		// integer digits past Precision−Scale is never storable (the insert is
-		// rejected), while a constant past it was measured to PROMOTE on the
-		// query path and compare mathematically — the refusal there is an
-		// accepted availability cost, taken because the integer widths' wrap
-		// behavior proves this class has no reliable single model (see
-		// decimalInPrecision — one story across the three sites). Scale
-		// truncation below cannot change integer digits, so gating
-		// pre-truncation is exact.
-		if !n.decimalInPrecision(a) || !n.decimalInPrecision(b) {
-			return 0, false
-		}
-		return compareCanonicalDecimals(truncateScale(a, n.Scale), truncateScale(b, n.Scale)), true
-	case NumericNone:
-		return 0, false // no storage model: refuse, fail closed
-	default:
-		return 0, false // future family nobody taught this switch: same refusal
-	}
-}
-
-// narrowFloat converts a canonical decimal form to the column's float domain
-// with a single correct rounding (ParseFloat at the exact bit width — never a
-// float64 detour, whose double rounding can land Float32 values one ULP off).
-// A magnitude the domain can't hold refuses the comparison: ClickHouse would
-// store ±Inf there, and matching infinities is a verdict this evaluator can't
-// prove cheaply, so the row is withheld — availability, never exposure. An
-// unknown width refuses too: ParseFloat silently treats any other bitSize as
-// 64, which would compare a Float32 column in the wrong (wider) domain — the
-// fail-open direction — instead of the zero-value-refuses contract the
-// integer and decimal families keep.
-func narrowFloat(canonical string, bits int) (float64, bool) {
-	if bits != 32 && bits != 64 {
-		return 0, false
-	}
-	f, err := strconv.ParseFloat(canonical, bits)
-	if err != nil {
-		return 0, false
-	}
-	return f, true
-}
-
-// truncateScale narrows a canonical decimal form to scale fractional digits,
-// truncating toward zero — ClickHouse's Decimal cast (1.005, 1.006 and 1.009
-// all store as 1.00 in a Decimal(10,2); rounding would predict 1.01). The
-// result is re-canonicalized (trailing zeros trimmed, bare "-0" folded) so it
-// stays valid compareCanonicalDecimals input.
-func truncateScale(canonical string, scale int) string {
-	intPart, frac, hasFrac := strings.Cut(canonical, ".")
-	if !hasFrac {
-		return canonical
-	}
-	if len(frac) > scale {
-		frac = frac[:scale]
-	}
-	for len(frac) > 0 && frac[len(frac)-1] == '0' {
-		frac = frac[:len(frac)-1]
-	}
-	if len(frac) > 0 {
-		return intPart + "." + frac
-	}
-	if intPart == "-0" {
-		return "0"
-	}
-	return intPart
-}
 
 // ColumnKind classifies a column's ClickHouse type for the in-memory row-filter
 // comparison. The zero value is ColumnOpaque, so a nil map, a column absent from
@@ -485,67 +274,5 @@ func compareScalar(rowVal any, filterVal string, spec ColumnSpec) (int, bool) {
 		return 0, false
 	default:
 		return 0, false // unknown future kind: refuse to compare, fail closed
-	}
-}
-
-// numericCanonical renders a payload value as the canonical decimal form the
-// numeric comparison consumes, ok=false for anything that is not a number a
-// ClickHouse numeric column could have stored: booleans, structured values and
-// null, spellings outside the JSON number grammar ("Inf", "NaN", "0x1f",
-// "007"), values whose exact digits were lost upstream (float64 at/past 2^53),
-// and anything past the canonical digit bound. String and json.Number inputs
-// take the claim side's own gates (CanonicalNumericLiteral / CanonicalScalar),
-// so the payload and constant sides can never disagree on what counts as a
-// number or how it is spelled.
-func numericCanonical(v any) (string, bool) {
-	switch x := v.(type) {
-	case string:
-		if len(x) > maxNumericOperandChars {
-			return "", false
-		}
-		return CanonicalNumericLiteral(x)
-	case json.Number:
-		if len(x) > maxNumericOperandChars {
-			return "", false
-		}
-		return CanonicalScalar(x)
-	case float64:
-		// CanonicalScalar applies the 2^53 exactness guard and renders
-		// positionally; the literal gate then re-canonicalizes the one
-		// rendering FormatFloat emits that canonical form forbids ("-0").
-		s, ok := CanonicalScalar(x)
-		if !ok {
-			return "", false
-		}
-		return CanonicalNumericLiteral(s)
-	default:
-		return "", false
-	}
-}
-
-// scalarString renders a JSON-decoded scalar as the exact BYTES compared under
-// ColumnText and ColumnOpaque — deliberately the payload's raw spelling, never
-// a canonical form: a String column stores the payload text verbatim, so byte
-// comparison against it must use that spelling (canonicalizing "1.0" to "1"
-// here would move equality away from what ClickHouse stores). Numeric coercion
-// deliberately does NOT live here — the ColumnNumeric arm routes both operands
-// through the claim side's canonical machinery (numericCanonical). Non-scalars
-// (arrays, objects, null) return ok=false so the predicate fails closed rather
-// than guessing. The float64 case serves callers that decoded without
-// UseNumber (the stream itself always does); -1 precision emits the shortest
-// round-trip form without an exponent, so integer IDs read back as "123", not
-// "1.23e+02".
-func scalarString(v any) (string, bool) {
-	switch x := v.(type) {
-	case string:
-		return x, true
-	case json.Number:
-		return string(x), true
-	case float64:
-		return strconv.FormatFloat(x, 'f', -1, 64), true
-	case bool:
-		return strconv.FormatBool(x), true
-	default:
-		return "", false
 	}
 }
